requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
dj-database-url==2.2.0
dj3-cloudinary-storage==0.0.6
Django==5.1.2
django-allauth==0.57.2
djangorestframework==3.15.2
gunicorn==23.0.0
oauthlib==3.2.2
psycopg2==2.9.10
PyJWT==2.9.0
python3-openid==3.2.0
requests-oauthlib==2.0.0
sqlparse==0.5.1
whitenoise==6.5.0
